--- examples/autoForm/react/index.ts
@@ -1,7 +1,0 @@
-export { AutoFormDemo } from "./AutoFormDemo";
-export { RenderPropDemo } from "./RenderPropDemo";
-export { ErrorComponentDemo } from "./ErrorComponentDemo";
-export { NestedArrayDemo } from "./NestedArrayDemo";
-export { SimpleNestedTest } from "./SimpleNestedTest";
-export { ApiComparison } from "./ApiComparison";
-
