+++ examples/react/index.ts
@@ -1,9 +1,8 @@
-export { AutoFormDemo } from "../autoForm/react/AutoFormDemo";
-export { RenderPropDemo } from "../autoForm/react/RenderPropDemo";
-export { ErrorComponentDemo } from "../autoForm/react/ErrorComponentDemo";
-export { NestedArrayDemo } from "../autoForm/react/NestedArrayDemo";
-export { SimpleNestedTest } from "../autoForm/react/SimpleNestedTest";
-export { ApiComparison } from "../autoForm/react/ApiComparison";
-export { UseFormDemo } from "../useForm/react/UseFormDemo";
-export { DirectArrayTest } from "../useForm/react/DirectArrayTest";
-
+export { AutoFormDemo } from "./autoForm/AutoFormDemo";
+export { RenderPropDemo } from "./autoForm/RenderPropDemo";
+export { ErrorComponentDemo } from "./autoForm/ErrorComponentDemo";
+export { NestedArrayDemo } from "./autoForm/NestedArrayDemo";
+export { SimpleNestedTest } from "./autoForm/SimpleNestedTest";
+export { ApiComparison } from "./autoForm/ApiComparison";
+export { UseFormDemo } from "./useForm/UseFormDemo";
+export { DirectArrayTest } from "./useForm/DirectArrayTest";
