--- examples/useForm/react/index.ts
@@ -1,3 +1,0 @@
-export { UseFormDemo } from "./UseFormDemo";
-export { DirectArrayTest } from "./DirectArrayTest";
-
